Refactor tree
